requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0

# Visualization
matplotlib>=3.7.0
plotly>=5.15.0
seaborn>=0.12.0
kaleido>=0.2.1

# Machine Learning
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
evaluate>=0.4.0

# API clients
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.7.0

# Hugging Face ecosystem
langchain-huggingface>=0.0.3
huggingface-hub>=0.16.0

# Progress bars and utilities
tqdm>=4.65.0
tenacity>=8.2.0

# Optional: For fine-tuning and advanced features
# Unsloth and TRL (uncomment if using open-source models)
# unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git
# trl>=0.7.0
# peft>=0.5.0
# accelerate>=0.21.0
# bitsandbytes>=0.41.0

# Development and testing (optional)
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Environment management
python-dotenv>=1.0.0

# System utilities
psutil>=5.9.0
requests>=2.31.0